+++ src/api/projects/create.rs
@@ -515,6 +515,9 @@ pub struct CreateProject<'a> {
     /// Whether to enable email notifications or not.
     #[builder(default)]
     emails_disabled: Option<bool>,
+    /// Whether the default set of award emojis are shown for this project.
+    #[builder(default)]
+    show_default_award_emojis: Option<bool>,
     /// Whether outdated diff discussions are resolved when a merge request is updated or not.
     #[builder(default)]
     resolve_outdated_diff_discussions: Option<bool>,
@@ -778,6 +781,7 @@ impl<'a> Endpoint for CreateProject<'a> {
             .push_opt("requirements_access_level", self.requirements_access_level)
             .push_opt("analytics_access_level", self.analytics_access_level)
             .push_opt("emails_disabled", self.emails_disabled)
+            .push_opt("show_default_award_emojis", self.show_default_award_emojis)
             .push_opt(
                 "resolve_outdated_diff_discussions",
                 self.resolve_outdated_diff_discussions,
@@ -1406,6 +1410,25 @@ mod tests {
         let endpoint = CreateProject::builder()
             .name("name")
             .emails_disabled(true)
+            .build()
+            .unwrap();
+        api::ignore(endpoint).query(&client).unwrap();
+    }
+
+    #[test]
+    fn endpoint_show_default_award_emojis() {
+        let endpoint = ExpectedUrl::builder()
+            .method(Method::POST)
+            .endpoint("projects")
+            .content_type("application/x-www-form-urlencoded")
+            .body_str(concat!("name=name", "&show_default_award_emojis=false"))
+            .build()
+            .unwrap();
+        let client = SingleTestClient::new_raw(endpoint, "");
+
+        let endpoint = CreateProject::builder()
+            .name("name")
+            .show_default_award_emojis(false)
             .build()
             .unwrap();
         api::ignore(endpoint).query(&client).unwrap();